requirements.txt
behave==1.2.6
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
flask==3.0.3
idna==3.8
importlib-metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
kafka==1.3.5
kafka-python==2.0.2
MarkupSafe==2.1.5
packaging==24.1
parse==1.20.2
parse-type==0.6.3
pluggy==1.5.0
prometheus-client==0.20.0
PyJWT==2.9.0
pytest==8.3.2
requests==2.32.3
six==1.16.0
tomli==2.0.1
urllib3==2.2.2
werkzeug==3.0.4
zipp==3.20.1
